fix eliminar campañas y acciones
ahora se eliminan acciones al eliminar campañas y acciones y  del tasker

+++ src/main/java/com/PubliciBot/DAO/Neodatis/UsuarioDAONeodatis.java
@@ -1,8 +1,7 @@
 package com.PubliciBot.DAO.Neodatis;
 
 import com.PubliciBot.DAO.Interfaces.UsuarioDAO;
-import com.PubliciBot.DM.Campana;
-import com.PubliciBot.DM.Usuario;
+import com.PubliciBot.DM.*;
 import org.neodatis.odb.ODB;
 import org.neodatis.odb.ODBFactory;
 import org.neodatis.odb.Objects;
@@ -10,6 +9,7 @@ import org.neodatis.odb.core.query.IQuery;
 import org.neodatis.odb.core.query.criteria.Where;
 import org.neodatis.odb.impl.core.query.criteria.CriteriaQuery;
 
+import java.util.ArrayList;
 import java.util.Collection;
 
 /**
@@ -30,17 +30,43 @@ public class UsuarioDAONeodatis extends DAONeodatis<Usuario> implements UsuarioD
             nuevo.getCampanas().putAll(user.getCampanas());
             ODB odb = null;
             try {
+
                 odb = ODBFactory.open(fileNameNeodatisDB);
                 IQuery usuarioCorrecto = new CriteriaQuery(Usuario.class, Where.equal("mail", user.getMail()));
                 Objects<Usuario> usuarioRecuperado = odb.getObjects(usuarioCorrecto);
                 recovered = usuarioRecuperado.getFirst();
+
                 Collection campanas = recovered.getCampanas().values();
-                for(Object obj: campanas){
-                    Campana c = (Campana) obj;
-                    odb.delete(c);
+                if(campanas.size()>0) {
+                    for (Object obj : campanas) {
+                        Campana c = (Campana) obj;
+                        ArrayList<AccionPublicitaria> acs = c.getAcciones();
+                        if (acs.size() > 0) {
+                            for (AccionPublicitaria ac : acs) {
+                                /*
+                                Medio medio = ac.getMedio();
+                                TipoMedio tp = medio.getTipoMedio();
+                                odb.delete(tp);
+                                odb.delete(medio);
+                                */
+                                odb.delete(ac);
+                            }
+                        }
+                        ArrayList<Post> posts = c.getPosts();
+                        if(posts.size()>0) {
+                            for (Post p : c.getPosts()) {
+                                odb.delete(p);
+                            }
+                        }
+
+
+                        odb.delete(c);
+                    }
                 }
+
                 odb.delete(recovered);
                 odb.store(nuevo);
+
             } catch (Exception e) {
                 e.printStackTrace();
             } finally {
